chore: fix date formatting
The "hugo-dusk" theme formats dates as "2 Jan, 2006" which has bothered
me for a while as the comma is unnecessary when the day and month are in
this order.

I've updated the partials to more consistently format the dates as
"2006-01-02" which matches the format the bottom of post and talk pages.